Extract components for library header
